// frontend/src/lib/generated/types.gen.ts
export enum DirectionEnum {
  LTR = "ltr",
  RTL = "rtl",
}

/**
 * An enum field
 */
export enum IconPositionEnum {
  BEFORE_TEXT = "before_text",
  AFTER_TEXT = "after_text",
}

/**
 * An enum field
 */
export enum AlignmentEnum {
  CENTER = "center",
  LANGUAGE_DIRECTION = "language-direction",
}

/**
 * An enumeration field
 */
export enum VariantEnum {
  TEXT_ABOVE_BACKGROUND = "text-above-background",
  TEXT_OVER_BACKGROUND = "text-over-background",
  TEXT_BELOW_BACKGROUND = "text-below-background",
}